--- lib/src/commands/update/update.dart
@@ -1,1 +1,0 @@
-export 'update_command.dart';
