+++ lib/src/data/generators/number_theme_extension_generator.dart
@@ -1,10 +1,11 @@
 import 'package:code_builder/code_builder.dart';
-import 'package:figmage/src/generators/theme_extension_generator.dart';
+import 'package:figmage/src/data/generators/values_by_mode_theme_extension_generator.dart';
 
 /// {@template number_theme_extension_generator}
-/// A [ThemeExtensionGenerator] for numbers from doubles.
+/// A [ValuesByModeThemeExtensionGenerator] for numbers from doubles.
 /// {@endtemplate}
-class NumberThemeExtensionGenerator extends ThemeExtensionGenerator<double> {
+class NumberThemeExtensionGenerator
+    extends ValuesByModeThemeExtensionGenerator<double> {
   /// {@macro number_theme_extension_generator}
   NumberThemeExtensionGenerator({
     required super.className,
