+++ lib/src/domain/generators/theme_class_generator.dart
@@ -1,0 +1,28 @@
+/// {@template generator}
+/// The superclass for all generators that can generate a dart file as a String
+/// output.
+///
+/// All generators generate a `BuildContext` extension.
+/// {@endtemplate}
+abstract interface class ThemeClassGenerator {
+  /// The prefix for all generated files that are generated by this package.
+  static const generatedFilePrefix = '''
+      // coverage:ignore-file
+      // GENERATED CODE - DO NOT MODIFY BY HAND
+      // ignore_for_file: type=lint
+  ''';
+
+  /// The name of the generated class.
+  String get className;
+
+  /// Whether the return type of the `BuildContext` extension should be
+  /// nullable.
+  ///
+  /// If this is false, the extension will contain a null assertion, which means
+  /// your code will throw if you haven't provided the respective
+  /// `ThemeExtension` in the widget tree.
+  bool get buildContextExtensionNullable;
+
+  /// Generates a String
+  String generate();
+}
