+++ test/src/data/generators/spacer_generator_test.dart
@@ -1,5 +1,5 @@
 import 'package:code_builder/code_builder.dart';
-import 'package:figmage/src/generators/spacer_generator.dart';
+import 'package:figmage/src/data/generators/spacer_generator.dart';
 import 'package:test/test.dart';
 
 void main() {
